requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-allauth==0.61.1
django-cors-headers==4.3.1
requests==2.31.0
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
Pillow==12.0.0
qrcode[pil]==8.2

# Development
black==24.1.1
flake8==7.0.0
